+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Fixed a task note opened with "Open as note" never opening in the task editor again
 - Fixed projects, statuses, and priorities with an invalid color breaking the color picker
+- Fixed tasks failing to save or losing their description after renaming their project ([#359](https://github.com/dotpm/obsidian-pm/issues/359))
 
 ## [2.5.2] - 2026-09-21